rename dqn agent file
